Add MPLWidgetCompatibility base class for MPL 1.0 compatibility.
* Subclass from `object` so Python properties work (for MPL 1.0 compatibility).
* Move pre-MPL 1.2 compatibily (i.e. AxesWidget) to MPLWidgetCompatibility.
* I'm not sure multiple inheritance is the best solution, but it works.
